SessionContext consumes ActivityContext in order to keep track of currentSection and CurrentModule
